requirements.txt
Mezzanine==3.1.10
docker-py
boto
django-storages
django-mama-cas
pyzmq
ipython[notebook]
